Refactor and modularize Pythagoras-related codebase
Refactored `PythagorasHandler`, `PythagorasClient`, and `PythagorasQuery` to improve modularity and maintainability. Removed obsolete methods and redundant code. Introduced helper classes (`PythagorasQueryHelpers`, `QueryStringWriter`, `OperatorMaps`) and new enums/records (`Op`, `QueryRequest`, `BuildingSearchResult`).

Updated test cases to use concise syntax and reflect changes. Consolidated functionality into separate files, improving readability and reducing duplication.
