+++ src/ume-app-estateservice/Umea.se.EstateService.ServiceAccess/Pythagoras/Enum/Op.cs
@@ -1,0 +1,8 @@
+﻿namespace Umea.se.EstateService.ServiceAccess.Pythagoras.Enum;
+
+public enum Op
+{
+    Eq, Ne, Gt, Ge, Lt, Le,
+    LikeExact, LikeAnywhere, LikeStarts, LikeEnds,
+    ILikeExact, ILikeAnywhere, ILikeStarts, ILikeEnds
+}
