+++ src/ume-app-estateservice/Umea.se.EstateService.ServiceAccess/Pythagoras/WorkspaceSearchResult.cs
@@ -1,16 +1,5 @@
 namespace Umea.se.EstateService.ServiceAccess.Pythagoras;
 
-public sealed class BuildingSearchResult
-{
-    public int Id { get; init; }
-
-    public Guid? Uid { get; init; }
-
-    public string Name { get; init; } = string.Empty;
-
-    public string? PopularName { get; init; }
-}
-
 public sealed class WorkspaceSearchResult
 {
     public int Id { get; init; }
