+++ src/ume-app-estateservice/Umea.se.EstateService.Test/Pythagoras/BuildingControllerTests.cs
@@ -25,7 +25,7 @@ public class BuildingControllerTests
 
         IReadOnlyList<BuildingModel> buildings = await controller.GetBuildingsAsync(CancellationToken.None);
 
-        buildings.Select(b => b.Id).ShouldBe(new[] { 1, 2 });
+        buildings.Select(b => b.Id).ShouldBe([1, 2]);
         client.LastQueryString.ShouldBe("maxResults=50");
         client.LastEndpoint.ShouldBe("rest/v1/building");
     }
