+++ src/ume-app-estateservice/Umea.se.EstateService.Test/Pythagoras/PythagorasQueryTests.cs
@@ -1,5 +1,6 @@
 using System.Text.Json.Serialization;
 using Umea.se.EstateService.ServiceAccess.Pythagoras.Api;
+using Umea.se.EstateService.ServiceAccess.Pythagoras.Enum;
 
 namespace Umea.se.EstateService.Test.Pythagoras;
